wc in separate file as requested
