+++ src/cowrie/commands/__init__.py
@@ -8,7 +8,9 @@ __all__ = [
     'base64',
     'busybox',
     'cat',
+    'chpasswd',
     'curl',
+    'crontab',
     'dd',
     'env',
     'ethtool',
@@ -32,6 +34,7 @@ __all__ = [
     'tar',
     'uname',
     'ulimit',
+    'wc',
     'wget',
     'which',
     'perl',
